refactor: update imports for new icon components
