+++ src/modules/tracking/components/AllSessionsDoneCounter.tsx
@@ -1,8 +1,8 @@
 import { SESSIONS_COUNTERS } from '@/presentation/globals/mocks/tracking';
-import { CardTitle } from '../../../presentation/modules/dashboard/card/components/CardTitle';
-import { Counter } from './Counter';
+import { CardTitle } from '@/presentation/modules/dashboard/card/components/CardTitle';
+import { Counter } from '@/modules/tracking/components/Counter';
 import { Card } from '@/presentation/modules/dashboard/card/components/Card';
-import { IconTrophy } from '@/presentation/globals/components/icons/Icons';
+import { IconTrophy } from '@/presentation/globals/components/icons/outline/IconTrophy';
 
 export function AllSessionsDoneCounter({ className }: { className?: string }) {
   const { totalSessions } = SESSIONS_COUNTERS;
