+++ src/modules/tracking/components/SessionsHistory.tsx
@@ -7,10 +7,8 @@ import { MonthDateSelector } from '@/presentation/modules/calendar/components/Mo
 import { YearDateSelector } from '@/presentation/modules/calendar/components/YearDateSelector';
 import { CalendarContext } from '@/presentation/modules/calendar/contexts/calendar-context';
 import { CardTitle } from '@/presentation/modules/dashboard/card/components/CardTitle';
-import {
-  IconCalendarClock,
-  IconClipboardList,
-} from '@/presentation/globals/components/icons/Icons';
+import { IconClipboardList } from '@/presentation/globals/components/icons/outline/IconClipboardList';
+import { IconCalendarClock } from '@/presentation/globals/components/icons/outline/IconCalendarClock';
 import { SubtleCard } from '@/presentation/modules/dashboard/card/components/SubtleCard';
 import { getSessionFromDate } from '@/presentation/globals/lib/dates';
 import { LastSessions } from '@/modules/session/presentation/ui/components/LastSessions';
