+++ src/presentation/modules/form/components/modal-form/ModalForm.tsx
@@ -1,7 +1,10 @@
-import { IconXMark } from '@/presentation/globals/components/icons/Icons';
-import { Button } from '../../../button/components/Button';
 import { FormProvider } from 'react-hook-form';
-import { type ModalFormHookProps, useFormHook } from '../../hooks/useFormHook';
+import { IconXMark } from '@/presentation/globals/components/icons/outline/IconXMark';
+import { Button } from '@/presentation/modules/button/components/Button';
+import {
+  type ModalFormHookProps,
+  useFormHook,
+} from '@/presentation/modules/form/hooks/useFormHook';
 import type { ZodSchema } from 'zod/v3';
 
 export function ModalForm<T extends ZodSchema<any, any>>({
