+++ src/presentation/modules/theme/components/ToggleTheme.tsx
@@ -3,7 +3,8 @@
 import { useTheme } from 'next-themes';
 import { THEME } from '@/presentation/modules/theme/constants/client';
 import { DefaultIcon } from '@/presentation/globals/components/icons/DefaultIcon';
-import { IconMoon, IconSun } from '@/presentation/globals/components/icons/Icons';
+import { IconMoon } from '@/presentation/globals/components/icons/outline/IconMoon';
+import { IconSun } from '@/presentation/globals/components/icons/outline/IconSun';
 import { Button } from '@/presentation/modules/button/components/Button';
 
 export function ToggleTheme({ className = '' }: { className?: string }) {
